Cambios otro compu

+++ car-repair-service/src/main/java/tingeso/carrepairservice/entities/DetailEntity.java
@@ -32,7 +32,7 @@ public class DetailEntity {
     private List<String> repairNames;
 
     //Lista con los amount de las reparaciones
-    private List<Double> repairAmounts;
+    private List<Integer> repairAmounts;
 
     //Lista con los nombres de los descuentos
     private List<String> discountNames;
